+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## [0.3.0] - Added IndefiniteRotation3DEffect to allow for Infinite widget rotations!
+* It is now possible to make any widget to roate indefinetly
+* Possibility to choose speed of rotations
+* Choose the axis on which the Indefinite Rotation should work (x and/or y)
+* Optioanlly set number of rotations
+* Option to allow the user to interact while the widget rotates
+
 ## [0.2.1] - Inital and End Position
 * Added initial and end position for the child Widget to be placed. End position only works when returnsInPlace is set to true.
 * Added the possibility to make the child Widget rotate indefintely. This is still under development though, and I DO NOT recommend using it at this stage.